Replace ShouldBe<T>(T, object) with ShouldBe<T>(T,T) to force type matches

+++ Source/EasyNetQ.Tests/AutoSubscriberTests/When_autosubscribing_with_subscription_configuration_action.cs
@@ -64,7 +64,7 @@ namespace EasyNetQ.Tests.AutoSubscriberTests
             subscriptionConfiguration.AutoDelete.ShouldBeTrue();
             subscriptionConfiguration.CancelOnHaFailover.ShouldBeTrue();
             subscriptionConfiguration.Expires.ShouldEqual(10);
-            subscriptionConfiguration.PrefetchCount.ShouldEqual(10);
+            subscriptionConfiguration.PrefetchCount.ShouldEqual((ushort)10);
             subscriptionConfiguration.Priority.ShouldEqual(10);
 
         }
