+++ Source/EasyNetQ.Tests/ConnectionString/ConnectionStringGrammarTests.cs
@@ -15,7 +15,7 @@ namespace EasyNetQ.Tests.ConnectionString
             var host = ConnectionStringGrammar.Host.Parse("my.host.com:1234");
 
             host.Host.ShouldEqual("my.host.com");
-            host.Port.ShouldEqual(1234);
+            host.Port.ShouldEqual((ushort)1234);
         }
 
         [Fact]
@@ -24,7 +24,7 @@ namespace EasyNetQ.Tests.ConnectionString
             var host = ConnectionStringGrammar.Host.Parse("my.host.com");
 
             host.Host.ShouldEqual("my.host.com");
-            host.Port.ShouldEqual(0);
+            host.Port.ShouldEqual((ushort)0);
         }
 
         [Fact]
@@ -34,11 +34,11 @@ namespace EasyNetQ.Tests.ConnectionString
 
             hosts.Count().ShouldEqual(3);
             hosts.ElementAt(0).Host.ShouldEqual("host.one");
-            hosts.ElementAt(0).Port.ShouldEqual(1001);
+            hosts.ElementAt(0).Port.ShouldEqual((ushort)1001);
             hosts.ElementAt(1).Host.ShouldEqual("host.two");
-            hosts.ElementAt(1).Port.ShouldEqual(1002);
+            hosts.ElementAt(1).Port.ShouldEqual((ushort)1002);
             hosts.ElementAt(2).Host.ShouldEqual("host.three");
-            hosts.ElementAt(2).Port.ShouldEqual(1003);
+            hosts.ElementAt(2).Port.ShouldEqual((ushort)1003);
         }
 
         [Fact]
