+++ Source/EasyNetQ.Tests/ConnectionString/ConnectionStringParserTests.cs
@@ -30,10 +30,10 @@ namespace EasyNetQ.Tests.ConnectionString
             connectionConfiguration.VirtualHost.ShouldEqual("Copa");
             connectionConfiguration.UserName.ShouldEqual("Copa");
             connectionConfiguration.Password.ShouldEqual("abc_xyz");
-            connectionConfiguration.Port.ShouldEqual(12345);
-            connectionConfiguration.RequestedHeartbeat.ShouldEqual(3);
-            connectionConfiguration.PrefetchCount.ShouldEqual(2);
-            connectionConfiguration.Timeout.ShouldEqual(12);
+            connectionConfiguration.Port.ShouldEqual((ushort)12345);
+            connectionConfiguration.RequestedHeartbeat.ShouldEqual((ushort)3);
+            connectionConfiguration.PrefetchCount.ShouldEqual((ushort)2);
+            connectionConfiguration.Timeout.ShouldEqual((ushort)12);
             connectionConfiguration.PublisherConfirms.ShouldBeTrue();
             connectionConfiguration.CancelOnHaFailover.ShouldBeTrue();
         }
@@ -44,7 +44,7 @@ namespace EasyNetQ.Tests.ConnectionString
             const string connectionStringWithTimeout = "host=localhost;timeout=13";
             var connectionConfiguration = connectionStringParser.Parse(connectionStringWithTimeout);
 
-            connectionConfiguration.Timeout.ShouldEqual(13);
+            connectionConfiguration.Timeout.ShouldEqual((ushort)13);
         }
 
         [Fact]
@@ -80,10 +80,10 @@ namespace EasyNetQ.Tests.ConnectionString
         {
             ConnectionConfiguration connectionConfiguration = connectionStringParser.Parse("" + spec.amqpUri);
 
-            connectionConfiguration.Port.ShouldEqual(spec.port);
+            connectionConfiguration.Port.ShouldEqual((ushort)spec.port);
             connectionConfiguration.AMQPConnectionString.ShouldEqual(spec.amqpUri);
             connectionConfiguration.Hosts.First().Host.ShouldEqual(spec.host);
-            connectionConfiguration.Hosts.First().Port.ShouldEqual(spec.port);
+            connectionConfiguration.Hosts.First().Port.ShouldEqual((ushort)spec.port);
         }
 
 // ReSharper disable UnusedMethodReturnValue.Local
@@ -101,7 +101,7 @@ namespace EasyNetQ.Tests.ConnectionString
         {
             ConnectionConfiguration connectionConfiguration = connectionStringParser.Parse("amqp=amqp://host/;port=123");
 
-            connectionConfiguration.Port.ShouldEqual(123);
+            connectionConfiguration.Port.ShouldEqual((ushort)123);
         }
 
         [Fact]
@@ -109,7 +109,7 @@ namespace EasyNetQ.Tests.ConnectionString
         {
             ConnectionConfiguration connectionConfiguration = connectionStringParser.Parse("amqp=amqp://host:1234/");
 
-            connectionConfiguration.Port.ShouldEqual(1234);
+            connectionConfiguration.Port.ShouldEqual((ushort)1234);
         }
 
         [Fact]
